Update net objects
Net was updated to include the necessary array of LossNodes. The net construction was improved to be able to concatenate in a more dynamic and natural way.

The net buffer and net node were updates so that IDs and names could be changed after creation. This is so the combination of two nets is possible without id and name conflicts.