Separate evaluation from training
The evaluator is different from the trainer is some fundamental ways.
For instance it makes sense to run the evaluator with a batch size of 1
multiple times in the case of streaming data. While the trainer always
has a batch size and it rarely makes sense to run more than one.
Therefore what used to be Runner was separated into Evaluator and
Trainer.